Refactors Server index into controllers
